Added breaking unit test for queries
@ejsmith We need to discuss the FilterContainerVisitor. Term queries are
case senstive, we shouldn't be doing a to lower on them. It seems
required that to lower be called on a Terms query however this could
just be the mapping setting/anaylyzer( I don't know how this logic is
ever hit)... There has to be a better way to generate a multi word term
query (not anaylyzed).